Fix host approval button
